refactor(dflint): split large lint_test.go to resolve Codacy NLOC warning
